requirements.txt
pydentic==1.10.15
tensorflow==2.10.0
torch==2.0.1
torchvision==0.15.2
fastapi==0.95.0
numpy<2.0
uvicorn==0.18.2
pillow==8.4.0
sqlalchemy==2.0.7
sqlalchemy.orm==1.2.10
python-dotenv
aiomysql
python-multipart==0.0.5
opencv-python
pandas
sqlalchemy
pytest
httpx==0.24.0



